+++ package.json
@@ -31,6 +31,7 @@
     "next-themes": "^0.4.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "sonner": "^2.0.7",
     "tailwind-merge": "^3.3.1",
     "tailwindcss-animate": "^1.0.7"
   },
@@ -44,6 +45,7 @@
     "eslint": "^9",
     "eslint-config-next": "15.3.4",
     "postcss": "^8.5.6",
+    "supabase": "^2.33.9",
     "tailwindcss": "^3.4.17",
     "tailwindcss-cli": "^0.1.2",
     "typescript": "^5"
