+++ src/lib/middleware.ts
@@ -1,0 +1,14 @@
+import { NextResponse } from 'next/server'
+import { createMiddlewareClient } from '@supabase/ssr'
+import type { NextRequest } from 'next/server'
+
+export async function middleware(req: NextRequest) {
+  const res = NextResponse.next()
+  const supabase = createMiddlewareClient({ req, res })
+  await supabase.auth.getSession()
+  return res
+}
+
+export const config = {
+  matcher: ['/((?!_next/static|_next/image|favicon.ico).*)'],
+}
