+++ supabase/functions/confirm-request/deno.json
@@ -1,0 +1,3 @@
+{
+  "imports": {}
+}
